move contoller and test class to task folder
