+++ src/controllers/tasks/active.officers.controller.ts
@@ -1,7 +1,7 @@
 import { Request, Response } from "express";
-import { Templates } from "../types/template.paths";
-import { TASK_LIST_PATH, urlParams } from "../types/page.urls";
-import { getUrlWithCompanyNumber } from "../utils/url";
+import { Templates } from "../../types/template.paths";
+import { TASK_LIST_PATH, urlParams } from "../../types/page.urls";
+import { getUrlWithCompanyNumber } from "../../utils/url";
 
 export const get = (req: Request, res: Response) => {
   const companyNumber = req.params[urlParams.PARAM_COMPANY_NUMBER];
