+++ test/controllers/tasks/active.officers.controller.unit.ts
@@ -1,9 +1,9 @@
-import mocks from "../mocks/all.middleware.mock";
+import mocks from "../../mocks/all.middleware.mock";
 import request from "supertest";
-import app from "../../src/app";
-import { ACTIVE_OFFICERS_PATH } from "../../src/types/page.urls";
-import { companyAuthenticationMiddleware } from "../../src/middleware/company.authentication.middleware";
-import { getUrlWithCompanyNumber } from "../../src/utils/url";
+import app from "../../../src/app";
+import { ACTIVE_OFFICERS_PATH } from "../../../src/types/page.urls";
+import { companyAuthenticationMiddleware } from "../../../src/middleware/company.authentication.middleware";
+import { getUrlWithCompanyNumber } from "../../../src/utils/url";
 
 jest.mock("../../src/middleware/company.authentication.middleware");
 jest.mock("../../src/utils/url");
